refactor: move helper dashboard function to hook
